fx quant: move more functions to utils
Summary:

No logic change, improving readability

Test Plan:

CI

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
